WIP [LibOS] Try checkpoint-and-restore using files
This is for migration of Gramine processes from one machine to another.
The idea is similar to CRIU (Checkpoint Restore in Userspace) in Linux.

This is work in progress. This will probably be never implemented, as
checkpointing via the existing `fork()` functionality of Gramine is
limited to a single thread. At the same time, most programming runtimes
like Golang, Java, and Ruby use thread pools and background threads,
which cannot be checkpointed using `fork()`; also search for e.g.
"golang fork" and you'll see that these runtimes do not even implement
fork (only fork + exec or posix_spawn).

Saving this commit purely to not lose my early attempts at adding this
functionality in Gramine itself (may be useful for trivial C/C++ apps).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>